Move token logic from controller to service
