+++ WebAPI/WebAPI/Helpers/Services/AccountService.cs
@@ -1,0 +1,39 @@
+﻿using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+using WebAPI.Interface.Services;
+
+namespace WebAPI.Helpers.Services;
+
+public class AccountService : IAccountService
+{
+    private readonly IConfiguration _configuration;
+
+    public AccountService(IConfiguration configuration)
+    {
+        _configuration = configuration;
+    }
+
+    public string CreateJwsToken(string email)
+    {
+        var tokenHandler = new JwtSecurityTokenHandler();
+        var key = Encoding.ASCII.GetBytes(_configuration["JwtSettings:Key"]!);
+        var tokenDescriptor = new SecurityTokenDescriptor
+        {
+            Subject = new ClaimsIdentity(new[]
+            {
+                new Claim(ClaimTypes.Name, email),
+            }),
+            Expires = DateTime.UtcNow.AddMonths(8),
+            SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature),
+            Issuer = "https://manero.com",
+            Audience = "https://manero.com"
+        };
+
+        var token = tokenHandler.CreateToken(tokenDescriptor);
+        var tokenString = tokenHandler.WriteToken(token);
+
+        return tokenString;
+    }
+}
