+++ WebAPI/WebAPI/Interface/Services/IAccountService.cs
@@ -1,0 +1,6 @@
+﻿namespace WebAPI.Interface.Services;
+
+public interface IAccountService
+{
+    string CreateJwsToken(string email);
+}
